Fix: Constants grouping
