+++ src/main/java/frc/robot/RobotContainer.java
@@ -4,8 +4,6 @@ import org.littletonrobotics.junction.networktables.LoggedDashboardChooser;
 
 import com.pathplanner.lib.auto.AutoBuilder;
 
-import edu.wpi.first.math.geometry.Pose2d;
-import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine;
@@ -17,8 +15,6 @@ import frc.robot.subsystems.Drivetrain.ModuleIOSim;
 import frc.robot.subsystems.Drivetrain.ModuleIOSpark;
 import frc.robot.subsystems.Drivetrain.SwerveDriveSubsystem;
 import frc.robot.commands.SystemIdCommands;
-import frc.robot.commands.auto.*;
-import frc.robot.constants.SimConstants;
 
 public class RobotContainer {
 
@@ -32,7 +28,7 @@ public class RobotContainer {
     private final LoggedDashboardChooser<Command> autoChooser;
 
     public RobotContainer() {
-        switch (SimConstants.currentMode) {
+        switch (Constants.currentMode) {
             // Instantiate input/output for their respective modes
             case REAL:
                 drivetrain = new SwerveDriveSubsystem(
