+++ src/main/java/frc/robot/commands/SwerveCommand.java
@@ -7,9 +7,9 @@ import java.util.function.Supplier;
 
 
 import edu.wpi.first.math.filter.SlewRateLimiter;
-import frc.robot.constants.SwerveDriveConstants.DriveConstants;
-import frc.robot.constants.RemoteConstants.OIConstants;
+import frc.robot.Constants.OIConstants;
 import frc.robot.subsystems.Drivetrain.SwerveDriveSubsystem;
+import frc.robot.subsystems.Drivetrain.SwerveDriveConstants.DriveConstants;
 import edu.wpi.first.math.kinematics.ChassisSpeeds;
 
 public class SwerveCommand extends Command {
