+++ src/main/java/frc/robot/commands/SystemIdCommands.java
@@ -7,8 +7,8 @@ import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
-import frc.robot.constants.SwerveDriveConstants.DriveConstants;
 import frc.robot.subsystems.Drivetrain.SwerveDriveSubsystem;
+import frc.robot.subsystems.Drivetrain.SwerveDriveConstants.DriveConstants;
 
 import java.text.DecimalFormat;
 import java.text.NumberFormat;
