--- src/main/java/frc/robot/constants/RemoteConstants.java
@@ -1,14 +1,0 @@
-package frc.robot.constants;
-
-public class RemoteConstants {
-    public static final class OIConstants {
-        public static final int kDriverJoystickPort = 0;
-
-        public static final int kDriverYAxis = 1;
-        public static final int kDriverXAxis = 0;
-        public static final int kDriverRotAxis = 4;
-        public static final int kDriverFieldOrientedButtonIdx = 1;
-
-        public static final double kDeadband = 0.1;
-    }
-}
