--- src/main/java/frc/robot/constants/SensorConstants.java
@@ -1,7 +1,0 @@
-package frc.robot.constants;
-
-public class SensorConstants {
-    public static final class DIO {
-    }
-    
-}
