+++ src/main/java/frc/robot/subsystems/Drivetrain/GyroPigeonIO.java
@@ -1,7 +1,5 @@
 package frc.robot.subsystems.Drivetrain;
 
-import static frc.robot.constants.SwerveDriveConstants.DriveConstants.*;
-
 import com.ctre.phoenix6.BaseStatusSignal;
 import com.ctre.phoenix6.StatusCode;
 import com.ctre.phoenix6.StatusSignal;
@@ -11,6 +9,8 @@ import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.util.Units;
 import edu.wpi.first.units.measure.Angle;
 import edu.wpi.first.units.measure.AngularVelocity;
+
+import static frc.robot.subsystems.Drivetrain.SwerveDriveConstants.DriveConstants.*;
 
 import java.util.Queue;
