+++ src/main/java/frc/robot/subsystems/Drivetrain/Module.java
@@ -13,13 +13,14 @@
 
 package frc.robot.subsystems.Drivetrain;
 
-import static frc.robot.constants.SwerveDriveConstants.DriveConstants.*;
-
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.kinematics.SwerveModulePosition;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
 import edu.wpi.first.wpilibj.Alert;
 import edu.wpi.first.wpilibj.Alert.AlertType;
+
+import static frc.robot.subsystems.Drivetrain.SwerveDriveConstants.DriveConstants.*;
+
 import org.littletonrobotics.junction.Logger;
 
 public class Module {
