+++ src/main/java/frc/robot/subsystems/Drivetrain/ModuleIOSpark.java
@@ -1,10 +1,9 @@
 package frc.robot.subsystems.Drivetrain;
 
-import static frc.robot.constants.SwerveDriveConstants.DriveConstants.*;
-import static frc.robot.constants.SwerveDriveConstants.ModuleConstants.*;
-import static frc.robot.constants.MotorConstants.*; 
-import static frc.robot.constants.SwerveDriveConstants.*;
-import static frc.robot.constants.SensorConstants.*;
+import static frc.robot.subsystems.Drivetrain.SwerveDriveConstants.MotorConstants.*;
+import static frc.robot.subsystems.Drivetrain.SwerveDriveConstants.*;
+import static frc.robot.subsystems.Drivetrain.SwerveDriveConstants.DriveConstants.*;
+import static frc.robot.subsystems.Drivetrain.SwerveDriveConstants.ModuleConstants.*;
 import static frc.lib.SparkUtil.*;
 
 import com.revrobotics.AbsoluteEncoder;
