+++ src/main/java/frc/robot/subsystems/Drivetrain/SwerveDriveConstants.java
@@ -1,8 +1,12 @@
-package frc.robot.constants;
+package frc.robot.subsystems.Drivetrain;
 
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.math.kinematics.SwerveDriveKinematics;
 import edu.wpi.first.math.util.Units;
+import edu.wpi.first.math.system.plant.DCMotor;
+
+import com.pathplanner.lib.config.ModuleConfig;
+import com.pathplanner.lib.config.RobotConfig;
 
 public class SwerveDriveConstants {
     public static final class ModuleConstants {
@@ -88,4 +92,40 @@ public class SwerveDriveConstants {
     public static final double turnSimD = 0.0;
     public static final double turnPIDMinInput = 0; // Radians
     public static final double turnPIDMaxInput = 2 * Math.PI; // Radians
+
+    public class MotorConstants {
+        // Drive motor configuration
+        public static final int driveMotorCurrentLimit = 50;
+        public static final DCMotor driveGearbox = DCMotor.getNEO(1);
+
+        // Turn motor configuration
+        public static final boolean turnInverted = false;
+        public static final int turnMotorCurrentLimit = 40;
+        public static final DCMotor turnGearbox = DCMotor.getNEO(1);
+    }
+    public class PathPlanConstants {
+        // PathPlanner configuration
+        public static final double wheelRadiusMeters = Units.inchesToMeters(2);
+        public static final double kPhysicalMaxSpeedMetersPerSecond = 4.47; // Free speed of NEO * kDriveEncoderRot2Meter
+        public static final int driveMotorCurrentLimit = 50;
+        public static final DCMotor driveGearbox = DCMotor.getNEO(1);
+    
+    
+        public static final double robotMassKg = 74.088;
+        public static final double robotMOI = 6.883;
+        public static final double wheelCOF = 1.0;
+        
+        public static final Translation2d[] moduleTranslations = TranslationConstants.moduleTranslations;
+        public static final RobotConfig robotConfig = new RobotConfig(
+                robotMassKg,
+                robotMOI,
+                new ModuleConfig(
+                        wheelRadiusMeters,
+                        kPhysicalMaxSpeedMetersPerSecond,
+                        wheelCOF,
+                        driveGearbox.withReduction(1 / ModuleConstants.kDriveMotorGearRatio),
+                        driveMotorCurrentLimit,
+                        1),
+                moduleTranslations);
+    }
 }
